chore: divide each components
